Improve scrape readability
The abridged scraper now writes a json file of the scrape. It finds
items that begin with "@." and turns them into json keys. The list
of items usually precede the numbered items in the format: A. B. ...
those are split up into list elements under each json key. The only
issue is that this format is not completely consistent so there are
some inconsistencies. As such there doesn't seem to be a nice clean
way to parse this. This implementation works somewhat well though.